Remove Exceptions package and refactor Album model to use ResourceNotFoundException

+++ src/main/java/com/codecool/spotify/controllers/ResourceNotFoundException.java
@@ -1,7 +1,7 @@
 package com.codecool.spotify.controllers;
 
 class ResourceNotFoundException extends RuntimeException{
-    ResourceNotFoundException(int id) {
+    ResourceNotFoundException(long id) {
         super("Could not find resource id: " + id);
     }
 }
